refactor: :recycle: rename file from `write_package_properties` to `write_properties`
