+++ src/seedcase_sprout/write_properties.py
